requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
archspec 
attrs==25.3.0
blinker==1.9.0
boltons 
Brotli 
cachetools==5.5.2
certifi 
cffi 
charset-normalizer
click==8.2.1
cryptography 
dataclasses-json==0.6.7
distro 
frozendict 
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna 
Jinja2==3.1.6
jiter==0.10.0
jsonpatch 
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.43
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
numpy==2.2.6
openai==1.82.1
orjson==3.10.18
packaging 
pandas==2.2.3
pip
pillow==11.2.1
platformdirs 
pluggy 
propcache==0.3.1
protobuf==4.25.3
pyarrow==20.0.0
pycosat 
pycparser 
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PySocks 
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests 
requests-toolbelt==1.0.0
rpds-py==0.25.1
ruamel.yaml 
setuptools==72.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
sqlite3 
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm 
truststore 
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3 
watchdog==6.0.0
wheel==0.43.0
yarl==1.20.0
zstandard==0.23.0

